feat: Moves submitting the action fact panel and adds social panel to its own file
