+++ src/components/logAction/ShareOnSocialPanel.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import useTranslation from '../customHooks/translations';
+
+const ShareOnSocialPanel = ({ quizAnswered, firstQuizAnswerCorrect }) => {
+  const translation = useTranslation();
+
+  //Update submitted action here and add quizAnswered and firstQuizAnswerCorrect
+
+  return <div>I dare you to match my action by...</div>;
+};
+
+export default ShareOnSocialPanel;
